Move static image renderer tests to orca test suite so that orca is installed on circleci
